+++ Modules/DataModel/Sources/PocketCastsDataModel/Private/Managers/EpisodeDataManager.swift
@@ -42,7 +42,8 @@ class EpisodeDataManager {
         "excludeFromEpisodeLimit",
         "starredModified",
         "deselectedChapters",
-        "deselectedChaptersModified"
+        "deselectedChaptersModified",
+        "metadata"
     ]
 
     // MARK: - Query
@@ -990,6 +991,7 @@ class EpisodeDataManager {
         values.append(episode.starredModified)
         values.append(DBUtils.nullIfNil(value: episode.deselectedChapters))
         values.append(episode.deselectedChaptersModified)
+        values.append(episode.rawMetadata as Any)
 
         if includeIdForWhere {
             values.append(episode.id)
@@ -1029,7 +1031,7 @@ extension EpisodeDataManager {
                     db.beginTransaction()
 
                     for episode in showInfo {
-                        try db.executeUpdate("INSERT OR REPLACE INTO EpisodeMetadata VALUES(?, ?);", values: [episode.key, episode.value])
+                        try db.executeUpdate("UPDATE \(DataManager.episodeTableName) SET metadata = ? WHERE uuid = ?;", values: [episode.value, episode.key])
                     }
 
                     db.commit()
@@ -1046,7 +1048,7 @@ extension EpisodeDataManager {
         return try await withCheckedThrowingContinuation { continuation in
             dbQueue.inDatabase { db in
                 do {
-                    let resultSet = try db.executeQuery("SELECT metadata from EpisodeMetadata WHERE episodeUuid = ?", values: [uuid])
+                    let resultSet = try db.executeQuery("SELECT metadata from \(DataManager.episodeTableName) WHERE uuid = ?", values: [uuid])
                     defer { resultSet.close() }
 
                     if resultSet.next(), let metadataData = resultSet.string(forColumn: "metadata")?.data(using: .utf8) {
@@ -1059,6 +1061,26 @@ extension EpisodeDataManager {
                     }
                 } catch {
                     FileLog.shared.addMessage("EpisodeDataManager.findEpisodeMetadata Episode metadata error: \(error)")
+                    continuation.resume(throwing: error)
+                }
+            }
+        }
+    }
+
+    func findRawEpisodeMetadata(uuid: String, dbQueue: FMDatabaseQueue) async throws -> String? {
+        return try await withCheckedThrowingContinuation { continuation in
+            dbQueue.inDatabase { db in
+                do {
+                    let resultSet = try db.executeQuery("SELECT metadata from \(DataManager.episodeTableName) WHERE uuid = ?", values: [uuid])
+                    defer { resultSet.close() }
+
+                    if resultSet.next() {
+                        continuation.resume(returning: resultSet.string(forColumn: "metadata"))
+                    } else {
+                        continuation.resume(returning: nil)
+                    }
+                } catch {
+                    FileLog.shared.addMessage("EpisodeDataManager.findRawEpisodeMetadata Episode metadata error: \(error)")
                     continuation.resume(throwing: error)
                 }
             }