+++ Modules/DataModel/Sources/PocketCastsDataModel/Public/Model/Episode+fromDatabase.swift
@@ -44,6 +44,7 @@ extension Episode {
         episode.starredModified = rs.longLongInt(forColumn: "starredModified")
         episode.deselectedChapters = rs.string(forColumn: "deselectedChapters")
         episode.deselectedChaptersModified = rs.longLongInt(forColumn: "deselectedChaptersModified")
+        episode.rawMetadata = rs.string(forColumn: "metadata")
         return episode
     }
 }